refactor: adjust font size for screen smaller than 425px
The design file provided has a mobile screen of 430px. I developed with
the spacing as close as possible to the design but the content overflows
if the screen is smaller than 430px. To fix this, I can use my own
judgement to adjust the sizing of overflowing element to make it
responsive but I still want to keep the look and feel of the original
sizing. Because of this, I choose another method that is to change the
default font size for screen smaller than 430px and 12 px work wonders.